refactor(web): extract DashboardKpis for reuse
