+++ apps/web/src/components/dashboard-kpis.tsx
@@ -1,0 +1,138 @@
+import { Card, Divider, Group, SimpleGrid, Skeleton, Stack, Tabs, Text, Title } from "@mantine/core"
+import { IconArrowDownRight, IconArrowUpRight, IconCash } from "@tabler/icons-react"
+import { useQuery } from "@tanstack/react-query"
+
+import {
+  getHairAssignedStatsForDate,
+  getHairAssignedThroughSaleStatsForDate,
+  getTransactionStatsForDate,
+  type StatCategory,
+} from "@/functions/dashboard"
+import { CURRENCIES } from "@/lib/currency"
+import { dashboardKeys } from "@/lib/query-keys"
+
+type StatTab = "total" | "average" | "count"
+
+export function DashboardKpis({ date, legalEntityId }: { date: string; legalEntityId: string }) {
+  const txQuery = useQuery({
+    queryKey: [...dashboardKeys.transactionStats(date), legalEntityId],
+    queryFn: () => getTransactionStatsForDate({ data: { date, legalEntityId: legalEntityId || undefined } }),
+  })
+  const hairAptQuery = useQuery({
+    queryKey: [...dashboardKeys.hairAssignedStats(date), legalEntityId],
+    queryFn: () => getHairAssignedStatsForDate({ data: { date, legalEntityId: legalEntityId || undefined } }),
+  })
+  const hairSaleQuery = useQuery({
+    queryKey: [...dashboardKeys.hairSaleStats(date), legalEntityId],
+    queryFn: () =>
+      getHairAssignedThroughSaleStatsForDate({ data: { date, legalEntityId: legalEntityId || undefined } }),
+  })
+
+  const txStats = txQuery.data
+  const hairApt = hairAptQuery.data
+  const hairSale = hairSaleQuery.data
+
+  return (
+    <Stack>
+      <SimpleGrid cols={{ base: 1, sm: 2, md: 3 }}>
+        {txStats
+          ? CURRENCIES.filter((c) => Number(txStats[c].count.current) > 0 || Number(txStats[c].count.previous) > 0).map(
+              (c) => <EnhancedStatCard key={c} title={`Transactions (${c})`} data={txStats[c]} />,
+            )
+          : Array.from({ length: 2 }).map((_, i) => <Skeleton key={i} h={240} />)}
+      </SimpleGrid>
+
+      <Title order={4}>Hair Assigned during Appointments</Title>
+      <SimpleGrid cols={{ base: 1, sm: 2, md: 4 }}>
+        {hairApt ? (
+          <>
+            <EnhancedStatCard title="Weight in Grams" data={hairApt.weightInGrams} />
+            <EnhancedStatCard title="Sold For" data={hairApt.soldFor} />
+            <EnhancedStatCard title="Profit" data={hairApt.profit} />
+            <EnhancedStatCard title="Price per Gram" data={hairApt.pricePerGram} defaultTab="average" />
+          </>
+        ) : (
+          Array.from({ length: 4 }).map((_, i) => <Skeleton key={i} h={240} />)
+        )}
+      </SimpleGrid>
+
+      <Title order={4}>Hair Assigned during Sale</Title>
+      <SimpleGrid cols={{ base: 1, sm: 2, md: 4 }}>
+        {hairSale ? (
+          <>
+            <EnhancedStatCard title="Weight in Grams" data={hairSale.weightInGrams} />
+            <EnhancedStatCard title="Sold For" data={hairSale.soldFor} />
+            <EnhancedStatCard title="Profit" data={hairSale.profit} />
+            <EnhancedStatCard title="Price per Gram" data={hairSale.pricePerGram} defaultTab="average" />
+          </>
+        ) : (
+          Array.from({ length: 4 }).map((_, i) => <Skeleton key={i} h={240} />)
+        )}
+      </SimpleGrid>
+    </Stack>
+  )
+}
+
+function EnhancedStatCard({
+  title,
+  data,
+  defaultTab = "total",
+}: {
+  title: string
+  data: StatCategory
+  defaultTab?: StatTab
+}) {
+  return (
+    <Card withBorder padding="md" radius="md">
+      <Group justify="space-between">
+        <Text size="xs" c="dimmed" tt="uppercase">
+          {title}
+        </Text>
+        <IconCash size={20} />
+      </Group>
+
+      <Tabs defaultValue={defaultTab} mt="md">
+        <Tabs.List>
+          <Tabs.Tab value="total">Total</Tabs.Tab>
+          <Tabs.Tab value="average">Average</Tabs.Tab>
+          <Tabs.Tab value="count">Count</Tabs.Tab>
+        </Tabs.List>
+        <StatTabPanel value="total" stat={data.total} />
+        <StatTabPanel value="average" stat={data.average} />
+        <StatTabPanel value="count" stat={data.count} />
+      </Tabs>
+
+      <Text fz="xs" c="dimmed" mt="sm">
+        Compared to previous period
+      </Text>
+    </Card>
+  )
+}
+
+function StatTabPanel({ value, stat }: { value: StatTab; stat: StatCategory[StatTab] }) {
+  const positive = stat.percentage >= 0
+  const Icon = positive ? IconArrowUpRight : IconArrowDownRight
+  const color = positive ? "teal" : "red"
+  return (
+    <Tabs.Panel value={value} pt="xs">
+      <Stack gap="xs">
+        <Group align="flex-end" gap="xs" mt="sm">
+          <Title order={3}>{stat.current}</Title>
+          <Group gap={4} c={color}>
+            <Text size="sm" fw={500}>
+              {stat.percentage}%
+            </Text>
+            <Icon size={16} />
+          </Group>
+        </Group>
+        <Text fz="xs" c="dimmed">
+          Previous: {stat.previous}
+        </Text>
+        <Divider my={4} />
+        <Text fz="xs" c="dimmed">
+          Difference: {stat.difference}
+        </Text>
+      </Stack>
+    </Tabs.Panel>
+  )
+}
